xhx_backend/requirements.txt
django==2.2.2
django-simpleui==4.0.2
django-filter==2.3.0
djangorestframework==3.11.0
uuid==1.30
pymongo==3.11.2
PyMySQL==0.9.3
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapy-redis==0.6.8
fake-useragent==0.1.11
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
redis-py-cluster==1.3.6
scrapy-redis-cluster ==0.4
django-redis==3.1.10
drf-extensions



